% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unique}
\alias{unique}
\title{Like the matlab function \code{unique(mat,'rows')}}
\usage{
unique(mat)
}
\arguments{
\item{mat}{The matrix to be considered}
}
\value{
The matrix without repeated rows, and ordered by first coloumn, second column...
}
\description{
Like the matlab function \code{unique(mat,'rows')}
}
\examples{
mat = matrix(c(1,2,1,3,4,2,6,4,8),3)
unique(mat)
}

